requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langgraph
chromadb
openai
google-generativeai
langchain-google-genai
pydantic
python-dotenv
tiktoken
langchain-openai
langchain-core
langchain-community
pypdf 
pymupdf
sentence-transformers
faiss-cpu
grandalf
beautifulsoup4
langchain-chroma
langsmith
fastapi-mcp
httpx
mcp
langchain-mcp-adapters
langgraph-checkpoint-sqlite
tavily-python
google-search-results